% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_categories.R
\name{assign_categories}
\alias{assign_categories}
\title{Assign Moose Stratification categories
Assign each block a high, medium or low category based on assigned values or quartile.}
\usage{
assign_categories(out_table, quartile = FALSE, low = 0.4, med = 0.7, high = 1)
}
\arguments{
\item{out_table}{data frame of moose stratification values}

\item{quartile}{TRUE or FALSE, if FALSE then categories will be based on low, med, high parameters}

\item{low}{numeric value to assign the highest value of the low category, default is 0.4}

\item{med}{numeric value to assign the highest value of the medium category, default is 0.7}

\item{high}{numeric value to assign the highest value of the high category, default is 1}
}
\value{
data table
}
\description{
Assign Moose Stratification categories
Assign each block a high, medium or low category based on assigned values or quartile.
}
\examples{
\dontrun{
assign_categories(out_table, quartile = FALSE, low = 40, med = 70, high = 100)
}
}
