% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_streams.R
\name{buffer_streams}
\alias{buffer_streams}
\title{Buffer streams by stream order}
\usage{
buffer_streams(streams, stream_order = c(3, 4, 5, 6, 7, 8), buffer_dist = 150)
}
\arguments{
\item{streams}{sf object with all streams}

\item{stream_order}{vector of numeric values with stream order, default =  c(3,4,5,6,7,8)}

\item{buffer_dist}{numeric buffer distance, default = 150}
}
\value{
sf object with dissolved and buffered stream network
}
\description{
Buffer streams by stream order
}
\examples{
\dontrun{
buffer_streams(streams, stream_order = c(3,4,5,6,7,8), buffer_dist = 150 )
}
}
