% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_areas.R
\name{calculate_areas}
\alias{calculate_areas}
\title{Calculate area of block}
\usage{
calculate_areas(blockno, aoi, out_dir)
}
\arguments{
\item{blockno}{list of block id numbers}

\item{aoi}{sf of aoi entire area with block ids}

\item{out_dir}{location of output folder}
}
\value{
table
}
\description{
Calculate area of block
}
\examples{
\dontrun{
calculate_area(blockno, in_aoi, out_path)
}
}
