% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_elev.R
\name{high_elevl}
\alias{high_elevl}
\title{Get base data required from bc data catalogue}
\usage{
high_elevl(dem, elev_thresh = 1500)
}
\arguments{
\item{dem}{an tif with elevation data}

\item{elev_thresh}{default elevation at which moose would be unlikely to pass, default = 1500m}
}
\value{
a geopackage with single polygon that shows elevation above the given threshold
}
\description{
Get base data required from bc data catalogue
}
\examples{
\dontrun{
high_elev(dem, 1500)
}
}
