% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_habit.R
\name{merge_habit}
\alias{merge_habit}
\title{Merge habitat areas}
\usage{
merge_habit(tmp_aoi, temp_out_dir)
}
\arguments{
\item{tmp_aoi}{sf object with aoi of interest}

\item{temp_out_dir}{file path where all prepared spatial files}
}
\value{
TRUE
}
\description{
Merge habitat areas
}
\examples{
\dontrun{
merge_nonhabitat(tmp_aoi, temp_out_dir)
}
}
