% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_nonhabit.R
\name{merge_nonhabit}
\alias{merge_nonhabit}
\title{Merge uninhabitable areas}
\usage{
merge_nonhabit(tmp_aoi, temp_out_dir)
}
\arguments{
\item{tmp_aoi}{sf object with aoi of interest}

\item{temp_out_dir}{file path where all prepared spatial files}
}
\value{
TRUE
}
\description{
Merge uninhabitable areas
}
\examples{
\dontrun{
merge_nonhabitat(tmp_aoi, temp_out_dir)
}
}
