% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skmoose-package.R
\docType{package}
\name{skmoose-package}
\alias{skmoose}
\alias{skmoose-package}
\title{skmoose: Stratifies Moose Survey Blocks}
\description{
Extract all required base data to assess moose inaccessible areas and mooose browse areas for a given area of interest. This package then outputs a table to assess high, medium and low predicted moose habitat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcperk/skmoose}
  \item \url{https://gcperk.github.io/skmoose/}
  \item \url{https://ninoxconsulting.github.io/skmoose/}
  \item Report bugs at \url{https://github.com/gcperk/skmoose/issues}
}

}
\keyword{internal}
