% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skmoose_example}
\alias{skmoose_example}
\title{helper function to get local file.path for test package dataset.}
\usage{
skmoose_example(path = NULL)
}
\arguments{
\item{path}{default location as NULL}
}
\value{
TRUE
}
\description{
helper function to get local file.path for test package dataset.
}
\examples{
\dontrun{
skmoose_example()
}
}
