% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steep_slope.R
\name{steep_slope}
\alias{steep_slope}
\title{Create steep slope exclusion layer}
\usage{
steep_slope(dem, slope_thresh = 55)
}
\arguments{
\item{dem}{elevation rast}

\item{slope_thresh}{a numeric threshold above which moose are unlikley to occur (degrees), default is 55.}
}
\value{
sf object with polygon rrepresenting higher than threshold area
}
\description{
Create steep slope exclusion layer
}
\examples{
\dontrun{
steep_slope(dem, 55)
}
}
