% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vri_browse.R
\name{vri_browse}
\alias{vri_browse}
\title{Filter Moose Browse Species}
\usage{
vri_browse(vri, species_codes = c("AT", "AC", "EP", "SB"))
}
\arguments{
\item{vri}{vri file as gpkg containing the fields SPECIES_CD_1}

\item{species_codes}{species codes to include. Default is c("AT", "AC","EP","SB")}
}
\value{
sf object with single polygon with deciduous browse
}
\description{
Filter Moose Browse Species
}
\examples{
vri_browse(vri, species_codes = c("AT", "AC","EP","SB"))

}
