% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skmoose-package.R
\docType{package}
\name{skmoose-package}
\alias{skmoose}
\alias{skmoose-package}
\title{skmoose: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gcperk/skmoose}
  \item Report bugs at \url{https://github.com/gcperk/skmoose/issues}
}

}
\keyword{internal}
